# back/docs/swagger.yaml
basePath: /
definitions:
  enum.ModelPortDirection:
    enum:
    - in
    - out
    type: string
    x-enum-varnames:
    - ModelPortDirectionIn
    - ModelPortDirectionOut
  enum.ModelType:
    enum:
    - atomic
    - coupled
    type: string
    x-enum-varnames:
    - Atomic
    - Coupled
  json.ModelColors:
    properties:
      bodyBackgroundColor:
        type: string
      headerBackgroundColor:
        type: string
      headerTextColor:
        type: string
    type: object
  json.ModelComponent:
    properties:
      instanceId:
        type: string
      instanceMetadata:
        $ref: '#/definitions/json.ModelMetadata'
      modelId:
        type: string
    required:
    - instanceId
    - modelId
    type: object
  json.ModelConnection:
    properties:
      from:
        $ref: '#/definitions/json.ModelLink'
      to:
        $ref: '#/definitions/json.ModelLink'
    required:
    - from
    - to
    type: object
  json.ModelLink:
    properties:
      instanceId:
        type: string
      port:
        type: string
    required:
    - instanceId
    - port
    type: object
  json.ModelMetadata:
    properties:
      alwaysShowExtraInfo:
        type: boolean
      alwaysShowToolbar:
        type: boolean
      backgroundColor:
        type: string
      modelColors:
        $ref: '#/definitions/json.ModelColors'
      parameters:
        items:
          $ref: '#/definitions/json.ModelParameter'
        type: array
      position:
        $ref: '#/definitions/json.ModelPosition'
      style:
        $ref: '#/definitions/json.ModelStyle'
      toolbarPosition:
        $ref: '#/definitions/json.ToolbarPosition'
      toolbarVisible:
        type: boolean
    required:
    - position
    - style
    type: object
  json.ModelParameter:
    properties:
      description:
        type: string
      name:
        type: string
      type:
        $ref: '#/definitions/json.ParameterType'
      value: {}
    required:
    - name
    - type
    - value
    type: object
  json.ModelPort:
    properties:
      id:
        type: string
      type:
        $ref: '#/definitions/enum.ModelPortDirection'
    required:
    - id
    - type
    type: object
  json.ModelPosition:
    properties:
      x:
        type: number
      "y":
        type: number
    required:
    - x
    - "y"
    type: object
  json.ModelStyle:
    properties:
      height:
        type: number
      width:
        type: number
    required:
    - height
    - width
    type: object
  json.ParameterType:
    enum:
    - int
    - float
    - bool
    - string
    - object
    type: string
    x-enum-varnames:
    - ParameterTypeInt
    - ParameterTypeFloat
    - ParameterTypeBool
    - ParameterTypeString
    - ParameterTypeObject
  json.ToolbarPosition:
    enum:
    - top
    - left
    - right
    - bottom
    type: string
    x-enum-varnames:
    - ToolbarPositionTop
    - ToolbarPositionLeft
    - ToolbarPositionRight
    - ToolbarPositionBottom
  model.Diagram:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      modelId:
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      updatedAt:
        type: string
      userId:
        type: string
      workspaceId:
        type: string
    required:
    - name
    type: object
  model.Library:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      models:
        items:
          $ref: '#/definitions/model.Model'
        type: array
      title:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  model.Model:
    properties:
      code:
        type: string
      components:
        items:
          $ref: '#/definitions/json.ModelComponent'
        type: array
      connections:
        items:
          $ref: '#/definitions/json.ModelConnection'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      libId:
        type: string
      metadata:
        $ref: '#/definitions/json.ModelMetadata'
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/json.ModelPort'
        type: array
      type:
        $ref: '#/definitions/enum.ModelType'
      updatedAt:
        type: string
      userId:
        type: string
    type: object
...
  request.DiagramRequest:
    properties:
      description:
        type: string
      name:
        type: string
      workspaceId:
        type: string
    required:
    - description
    - name
    - workspaceId
    type: object
...
  request.LibraryRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - description
    - title
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: This is the API documentation for Easy DEVS.
  title: Easy DEVS API
  version: "1.0"
paths:
...
  /diagram:
    get:
      description: Retrieve a list of all diagrams
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Diagram'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all diagrams
      tags:
      - diagrams
    post:
      consumes:
      - application/json
      description: Create a new diagram entry
      parameters:
      - description: Diagram data
        in: body
        name: diagram
        required: true
        schema:
          $ref: '#/definitions/request.DiagramRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Diagram'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a diagram
      tags:
      - diagrams
  /diagram/{id}:
    delete:
      description: Delete a diagram by its ID
      parameters:
      - description: Diagram ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a diagram
      tags:
      - diagrams
    get:
      description: Retrieve a single diagram by its ID
      parameters:
      - description: Diagram ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Diagram'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a diagram by ID
      tags:
      - diagrams
    patch:
      consumes:
      - application/json
      description: Update an existing diagram with partial data
      parameters:
      - description: Diagram ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/request.DiagramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Diagram'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Update a diagram
      tags:
      - diagrams
...
  /library:
    get:
      description: Retrieve a list of all libraries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Library'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all libraries
      tags:
      - libraries
    post:
      consumes:
      - application/json
      description: Create a new library entry
      parameters:
      - description: Library data
        in: body
        name: library
        required: true
        schema:
          $ref: '#/definitions/request.LibraryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Library'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a library
      tags:
      - libraries
  /library/{id}:
    delete:
      description: Delete a library by its ID
      parameters:
      - description: Library ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a library
      tags:
      - libraries
    get:
      description: Retrieve a single library by its ID
      parameters:
      - description: Library ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Library'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a library by ID
      tags:
      - libraries
    patch:
      consumes:
      - application/json
      description: Update an existing library with partial data
      parameters:
      - description: Library ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/request.LibraryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Library'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Update a library
      tags:
      - libraries
...
swagger: "2.0"